+++ src/parsers/number.ts
@@ -3,7 +3,7 @@ import { ParseFn, ParserError, ParserInput } from "./index";
 
 const aNumber: ParseFn<number> = (x: ParserInput): number => {
   if (isNil(x)) {
-    throw new ParserError("number", x);
+    throw new ParserError("number", `${x}`);
   }
 
   if (typeof x === "number") {