+++ src/parsers/withObject.ts
@@ -7,14 +7,14 @@ import { isNil } from "../util";
  * @param fn
  * @returns
  */
-export default function withObject<T>(fn: (object: Record<string, ParserInput>) => T): ParseFn<T> {
-  return (x) => {
+export default function withObject<T, O = Record<string, ParserInput>>(fn: (object: O) => T): ParseFn<T> {
+  return (x: O) => {
     if (isNil(x)) {
-      throw new ParserError("Object", x);
+      throw new ParserError("Object", typeof x);
     }
 
     if (Object.prototype.toString.call(x) !== "[object Object]") {
-      throw new ParserError("Object", x);
+      throw new ParserError("Object", typeof x);
     }
 
     return fn(x);