+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.1
+- Remove `any` type and use `unknown` wherever possible
+- Improve `withObject` and `fromObject` type signatures
+
 ## 2.1.0
 
 - Add `aNull` parser, allowing to parse `null`